Fix 403 error
Problem:
After spinning up the containers, attempting to access the page gives a various 403 errors.

Cause:
The page is trying to access bower_components, which is a symbolic link to ../../node_modules. This symbolic link navigates outside of the mounted volume, so is innacessible to the container.

Fix:
I modified the volume to mount the entire project instead of just the web folder, and added a symbolic link from html to web, since the webserver wants the web root to be html, and not web.